Feature/communications

+++ src/gameObjects/bullet.py
@@ -5,9 +5,10 @@ import pymunk
 from config import config
 from gameObjects.game_object import GameObject, IDCounter
 
+
 class Bullet(GameObject):
     def __init__(
-        self, space: pymunk.Space, coord: tuple[int, int], velocity: tuple[int, int]
+        self, space: pymunk.Space, coord: tuple[int, int], velocity: tuple[float, float]
     ):
         super().__init__(space, coord, velocity)
