+++ src/gameObjects/game_object.py
@@ -1,4 +1,5 @@
 from __future__ import annotations
+
 from collections import defaultdict
 
 import pymunk
@@ -20,7 +21,7 @@ class GameObject:
         self,
         space: pymunk.Space,
         coord: tuple[int, int],
-        velocity: tuple[int, int] = (0, 0),
+        velocity: tuple[float, float] = (0, 0),
         body_type=pymunk.Body.DYNAMIC,
     ):
         self.space = space
